Initial commit
10-07-2021 01-15p

Additionally .gitignore files were added, and unnecessary files were ignored.

The py.exe file was not ignored, as it is a project which runs a c++ file and is currently is a very basic or beta stage. It makes is easier to run or create a py file. It will be updated to a full version in the future.